feat(common): leverage trame-common for utility class/function
__New Features__:
- Add dependency to trame-common for providing generic classes
  that can be relevant across trame packages. This includes
  assets handling, execution helper (throttle, async task),
  decorators and abstract classes for building components or apps.
- Expose new TrameApp and TrameComponent classes from trame.app
  to simplify the creation of trame pieces when using inheritance
  instead of the @TrameApp() decorator.

__Deprecation and cleanup__
- trame.app.dev should be deprecated/removed but for now we just
  remove a method that has been broken for quite some time.
  In general that package is fairly advanced and internal.
  It should not be needed by 99.99% of folks.
- trame.app.jupyter has been removed since the Jupyter integration
  is directly part of the layouts and that package should not be
  used anymore.
