+++ pyproject.toml
@@ -6,9 +6,10 @@ authors = [
     {name = "Kitware Inc."},
 ]
 dependencies = [
-    "trame-server>=3.2.3,<4",
-    "trame-client>=3.4,<4",
-    "wslink>=2.1.3",
+    "trame-server>=3.4,<4",
+    "trame-client>=3.7.1,<4",
+    "trame-common>=0.2,<1",
+    "wslink>=2.3.3",
     "pyyaml", # for trame.tools.widgets
 ]
 requires-python = ">=3.9"
