+++ trame/app/testing.py
@@ -1,5 +1,5 @@
-import os
 import json
+import os
 
 
 class TrameServerMonitor:
