+++ trame/assets/mimetypes.py
@@ -1,0 +1,6 @@
+from trame_common.assets.mimetypes import add_mimetype_override, to_mime
+
+__all__ = [
+    "add_mimetype_override",
+    "to_mime",
+]
