+++ trame/decorators.py
@@ -1,0 +1,19 @@
+from trame_common.decorators.hot_reload import hot_reload
+from trame_common.decorators.klass import (
+    TrameApp,
+    change,
+    controller,
+    life_cycle,
+    trigger,
+)
+from trame_common.exec.asynchronous import task
+
+__all__ = [
+    "TrameApp",
+    "change",
+    "trigger",
+    "controller",
+    "life_cycle",
+    "hot_reload",
+    "task",
+]
