--- trame/decorators/__init__.py
@@ -1,17 +1,0 @@
-from .dev import hot_reload
-from .klass import (
-    TrameApp,
-    change,
-    trigger,
-    controller,
-    life_cycle,
-)
-
-__all__ = [
-    "TrameApp",
-    "change",
-    "trigger",
-    "controller",
-    "life_cycle",
-    "hot_reload",
-]
