--- trame/decorators/dev.py
@@ -1,5 +1,0 @@
-from trame_server.utils.hot_reload import hot_reload
-
-__all__ = [
-    "hot_reload",
-]
