+++ trame/tools/widgets/__main__.py
@@ -1,5 +1,6 @@
-import asyncio
 import argparse
+import asyncio
+
 from .generator import generate_trame_package
