+++ trame/utils/version.py
@@ -1,0 +1,5 @@
+from trame_common.utils.version import get_version
+
+__all__ = [
+    "get_version",
+]
